requirements.txt
flask
flask-cors
duckduckgo-search
langchain==0.1.16
wikipedia
langchain-community
langchain-openai
python-dotenv
pydantic
langchain-groq
ddgs
